apps/ai-api/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
pydantic-settings>=2.6.0,<3
httpx>=0.27.0,<1
python-multipart>=0.0.9,<1
mcp>=1.27.0,<2
